Add BRDF functions for Cook-Torrance model; refactor material module to include new brdf module
