+++ src/lib.rs
@@ -7,8 +7,8 @@ pub mod hittable;
 pub mod hittable_list;
 pub mod light;
 pub mod material;
-pub mod ray;
 pub mod primitives;
+pub mod ray;
 pub mod tracer;
 pub mod vec3;
 pub mod world;
