+++ src/material/brdf.rs
@@ -1,0 +1,66 @@
+use crate::color::Color;
+use crate::common;
+use crate::vec3;
+use crate::vec3::Vec3;
+
+pub fn fresnel_schlick(cos_theta: f32, f0: Color) -> Color {
+    f0 + (Color::new(1.0, 1.0, 1.0) - f0) * f32::powf(1.0 - cos_theta, 5.0)
+}
+
+pub fn geometry_schlick_ggx(n_dot: f32, roughness: f32) -> f32 {
+    let k = (roughness + 1.0).powi(2) / 8.0;
+    n_dot / (n_dot * (1.0 - k) + k)
+}
+
+pub fn sample_vndf_ggx(view: Vec3, roughness: f32) -> Vec3 {
+    // Transform view direction to hemisphere aligned with normal (Z+)
+    let v = vec3::unit_vector(Vec3::new(
+        roughness * view.x(),
+        roughness * view.y(),
+        view.z(),
+    ));
+
+    // Generate 2D random numbers
+    let (u1, u2) = common::random2();
+
+    // Construct orthonormal basis
+    let lensq = v.x() * v.x() + v.y() * v.y();
+    let (t1, t2) = if lensq > 0.0 {
+        let inv_len = 1.0 / lensq.sqrt();
+        (
+            Vec3::new(-v.y() * inv_len, v.x() * inv_len, 0.0),
+            Vec3::new(
+                -v.z() * v.x() * inv_len,
+                -v.z() * v.y() * inv_len,
+                lensq * inv_len,
+            ),
+        )
+    } else {
+        // view is aligned with z-axis
+        (Vec3::new(1.0, 0.0, 0.0), Vec3::new(0.0, 1.0, 0.0))
+    };
+
+    // Sample point on hemisphere
+    let r = u1.sqrt();
+    let phi = 2.0 * std::f32::consts::PI * u2;
+    let t1_coeff = r * phi.cos();
+    let t2_coeff = r * phi.sin();
+    let _s = 0.5 * (1.0 + v.z());
+    let t3 = (1.0 - u1).sqrt();
+
+    let h = t1 * t1_coeff + t2 * t2_coeff + v * t3;
+    vec3::unit_vector(Vec3::new(
+        roughness * h.x(),
+        roughness * h.y(),
+        h.z().max(1e-6),
+    ))
+}
+
+pub fn pdf_vndf_ggx(view: Vec3, half: Vec3, normal: Vec3, roughness: f32) -> f32 {
+    let a2 = roughness * roughness;
+    let n_dot_h = vec3::dot(normal, half).max(1e-6);
+    let v_dot_h = vec3::dot(view, half).max(1e-6);
+
+    let d = a2 / (std::f32::consts::PI * (n_dot_h * n_dot_h * (a2 - 1.0) + 1.0).powi(2));
+    d * n_dot_h / (4.0 * v_dot_h)
+}
